Changed uint to int. Added UpdateHtmlDataInAllImagesById, UpdateHtmlDataInFirstImagesByProductId and UpdateHtmlDataInFirstAndAllImagesByProductId.

+++ MOSTComputers.Services.ProductRegister/Services/ProductImageService.cs
@@ -9,8 +9,7 @@ using MOSTComputers.Models.Product.Models.Validation;
 using MOSTComputers.Models.Product.Models.Requests.ProductImage;
 using MOSTComputers.Services.ProductRegister.Mapping;
 using static MOSTComputers.Services.ProductRegister.Validation.CommonElements;
-using MOSTComputers.Models.Product.Models.Requests.ProductImageFileNameInfo;
-using System.Reflection.Metadata.Ecma335;
+using MOSTComputers.Utils.OneOf;
 
 namespace MOSTComputers.Services.ProductRegister.Services;
 
@@ -39,8 +38,10 @@ internal sealed class ProductImageService : IProductImageService
     private readonly IValidator<ServiceProductFirstImageCreateRequest>? _firstImageCreateRequestValidator;
     private readonly IValidator<ServiceProductFirstImageUpdateRequest>? _firstImageUpdateRequestValidator;
 
-    public IEnumerable<ProductImage> GetAllInProduct(uint productId)
+    public IEnumerable<ProductImage> GetAllInProduct(int productId)
     {
+        if (productId <= 0) return Enumerable.Empty<ProductImage>();
+
         return _productImageRepository.GetAllInProduct(productId);
     }
 
@@ -49,22 +50,28 @@ internal sealed class ProductImageService : IProductImageService
         return _productImageRepository.GetAllFirstImagesForAllProducts();
     }
 
-    public IEnumerable<ProductImage> GetAllFirstImagesForSelectionOfProducts(List<uint> productIds)
+    public IEnumerable<ProductImage> GetAllFirstImagesForSelectionOfProducts(List<int> productIds)
     {
+        productIds = RemoveValuesSmallerThanOne(productIds);
+
         return _productImageRepository.GetFirstImagesForSelectionOfProducts(productIds);
     }
 
-    public ProductImage? GetByIdInAllImages(uint id)
+    public ProductImage? GetByIdInAllImages(int id)
     {
+        if (id <= 0) return null;
+
         return _productImageRepository.GetByIdInAllImages(id);
     }
 
-    public ProductImage? GetFirstImageForProduct(uint productId)
+    public ProductImage? GetFirstImageForProduct(int productId)
     {
+        if (productId <= 0) return null;
+
         return _productImageRepository.GetByProductIdInFirstImages(productId);
     }
 
-    public OneOf<uint, ValidationResult, UnexpectedFailureResult> InsertInAllImages(ServiceProductImageCreateRequest createRequest,
+    public OneOf<int, ValidationResult, UnexpectedFailureResult> InsertInAllImages(ServiceProductImageCreateRequest createRequest,
         IValidator<ServiceProductImageCreateRequest>? validator = null)
     {
         ValidationResult validationResult = ValidateTwoValidatorsDefault(createRequest, validator, _createRequestValidator);
@@ -75,17 +82,17 @@ internal sealed class ProductImageService : IProductImageService
 
         createRequestInternal.DateModified = DateTime.Today;
 
-        OneOf<uint, UnexpectedFailureResult> result = _productImageRepository.InsertInAllImages(createRequestInternal);
+        OneOf<int, UnexpectedFailureResult> result = _productImageRepository.InsertInAllImages(createRequestInternal);
 
-        return result.Match<OneOf<uint, ValidationResult, UnexpectedFailureResult>>(
+        return result.Match<OneOf<int, ValidationResult, UnexpectedFailureResult>>(
             id => id, unexpectedFailure => unexpectedFailure);
     }
 
-    public OneOf<uint, ValidationResult, UnexpectedFailureResult> InsertInAllImagesAndImageFileNameInfos(ServiceProductImageCreateRequest createRequest,
-        uint? displayOrder = null,
+    public OneOf<int, ValidationResult, UnexpectedFailureResult> InsertInAllImagesAndImageFileNameInfos(ServiceProductImageCreateRequest createRequest,
+        int? displayOrder = null,
         IValidator<ServiceProductImageCreateRequest>? validator = null)
     {
-        if (displayOrder == 0)
+        if (displayOrder <= 0)
         {
             ValidationResult displayOrderResult = new();
 
@@ -102,9 +109,9 @@ internal sealed class ProductImageService : IProductImageService
 
         createRequestInternal.DateModified = DateTime.Today;
 
-        OneOf<uint, UnexpectedFailureResult> result = _productImageRepository.InsertInAllImagesAndImageFileNameInfos(createRequestInternal, displayOrder);
+        OneOf<int, UnexpectedFailureResult> result = _productImageRepository.InsertInAllImagesAndImageFileNameInfos(createRequestInternal, displayOrder);
 
-        return result.Match<OneOf<uint, ValidationResult, UnexpectedFailureResult>>(
+        return result.Match<OneOf<int, ValidationResult, UnexpectedFailureResult>>(
             id => id,
             unexpectedFailureResult => unexpectedFailureResult);
     }
@@ -160,23 +167,94 @@ internal sealed class ProductImageService : IProductImageService
             success => success, unexpectedFailure => unexpectedFailure);
     }
 
-    public bool DeleteInAllImagesById(uint id)
+    public OneOf<Success, ValidationResult, UnexpectedFailureResult> UpdateHtmlDataInAllImagesById(int imageId, string htmlData)
     {
+        ValidationResult validationResult = new();
+
+        if (imageId <= 0)
+        {
+            validationResult.Errors.Add(new(nameof(imageId), "Invalid id"));
+        }
+
+        if (string.IsNullOrEmpty(htmlData))
+        {
+            validationResult.Errors.Add(new(nameof(htmlData), "Cannot be null, empty, or whitespace"));
+        }
+
+        if (!validationResult.IsValid) return validationResult;
+
+        OneOf<Success, UnexpectedFailureResult> updateHtmlDataInAllImagesResult = _productImageRepository.UpdateHtmlDataInAllImagesById(imageId, htmlData);
+
+        return updateHtmlDataInAllImagesResult.Map<Success, ValidationResult, UnexpectedFailureResult>();
+    }
+
+    public OneOf<Success, ValidationResult, UnexpectedFailureResult> UpdateHtmlDataInFirstImagesByProductId(int productId, string htmlData)
+    {
+        ValidationResult validationResult = new();
+
+        if (productId <= 0)
+        {
+            validationResult.Errors.Add(new(nameof(productId), "Invalid id"));
+        }
+
+        if (string.IsNullOrEmpty(htmlData))
+        {
+            validationResult.Errors.Add(new(nameof(htmlData), "Cannot be null, empty, or whitespace"));
+        }
+
+        if (!validationResult.IsValid) return validationResult;
+
+        OneOf<Success, UnexpectedFailureResult> updateHtmlDataInFirstImagesResult = _productImageRepository.UpdateHtmlDataInFirstImagesByProductId(productId, htmlData);
+
+        return updateHtmlDataInFirstImagesResult.Map<Success, ValidationResult, UnexpectedFailureResult>();
+    }
+
+    public OneOf<Success, ValidationResult, UnexpectedFailureResult> UpdateHtmlDataInFirstAndAllImagesByProductId(int productId, string htmlData)
+    {
+        ValidationResult validationResult = new();
+
+        if (productId <= 0)
+        {
+            validationResult.Errors.Add(new(nameof(productId), "Invalid id"));
+        }
+
+        if (string.IsNullOrEmpty(htmlData))
+        {
+            validationResult.Errors.Add(new(nameof(htmlData), "Cannot be null, empty, or whitespace"));
+        }
+
+        if (!validationResult.IsValid) return validationResult;
+
+        OneOf<Success, UnexpectedFailureResult> updateHtmlDataInAllAndFirstImagesResult = _productImageRepository.UpdateHtmlDataInFirstAndAllImagesByProductId(productId, htmlData);
+
+        return updateHtmlDataInAllAndFirstImagesResult.Map<Success, ValidationResult, UnexpectedFailureResult>();
+    }
+
+    public bool DeleteInAllImagesById(int id)
+    {
+        if (id <= 0) return false;
+
         return _productImageRepository.DeleteInAllImagesById(id);
     }
 
-    public bool DeleteInAllImagesAndImageFilePathInfosById(uint id)
+    public bool DeleteInAllImagesAndImageFilePathInfosById(int id)
     {
+        if (id <= 0) return false;
+
         return _productImageRepository.DeleteInAllImagesAndImageFilePathInfosById(id);
     }
 
-    public bool DeleteAllImagesForProduct(uint productId)
+    public bool DeleteAllImagesForProduct(int productId)
     {
+        if (productId <= 0) return false;
+
         return _productImageRepository.DeleteAllWithSameProductIdInAllImages(productId);
     }
 
-    public bool DeleteInFirstImagesByProductId(uint productId)
+    public bool DeleteInFirstImagesByProductId(int productId)
     {
+        if (productId <= 0) return false;
+
         return _productImageRepository.DeleteInFirstImagesByProductId(productId);
     }
 }
